webui/requirements.txt
bs4
email_validator
Flask
Flask-Bootstrap
Flask-Login
Flask-SQLAlchemy
Flask-WTF
PyYAML
